Sanitise how we detect whether a user is on the mentor or mentee tab
