+++ metaswitch_tinder/pages/mentee_request.py
@@ -8,7 +8,7 @@ import metaswitch_tinder.database.matches
 from metaswitch_tinder.app import app
 from metaswitch_tinder import database
 from metaswitch_tinder.app_structure import href
-from metaswitch_tinder.components.auth import wait_for_login, current_username
+from metaswitch_tinder.components.session import wait_for_login, current_username
 from metaswitch_tinder.components.grid import create_equal_row
 from metaswitch_tinder.components.inputs import multi_dropdown_with_tags
