+++ metaswitch_tinder/pages/signin.py
@@ -7,7 +7,7 @@ from dash.dependencies import Output, State, Event
 from metaswitch_tinder import database
 from metaswitch_tinder.app import app
 from metaswitch_tinder.app_structure import href
-from metaswitch_tinder.components.auth import set_current_usename
+from metaswitch_tinder.components.session import set_current_usename
 from metaswitch_tinder.components.grid import create_equal_row
